Monkey-patch node:diagnostics_channel for fastify to allow tests under Node 18
