+++ .github/workflows/cibuild.yml
@@ -18,7 +18,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        nodeVersion: [ 20, 22 ]
+        nodeVersion: [ 18, 20, 22 ]
     name: Run Tests - Node ${{ matrix.nodeVersion }}
     steps:
       - uses: actions/checkout@v4
